+++ src/client/mod.rs
@@ -316,6 +316,28 @@ impl Writer for Context {
     }
 }
 
+#[cfg(any(feature = "rtu", feature = "tcp"))]
+pub(crate) async fn disconnect_framed<T, C>(
+    framed: tokio_util::codec::Framed<T, C>,
+) -> std::io::Result<()>
+where
+    T: tokio::io::AsyncWrite + Unpin,
+{
+    use tokio::io::AsyncWriteExt as _;
+
+    framed
+        .into_inner()
+        .shutdown()
+        .await
+        .or_else(|err| match err.kind() {
+            std::io::ErrorKind::NotConnected | std::io::ErrorKind::BrokenPipe => {
+                // Already disconnected.
+                Ok(())
+            }
+            _ => Err(err),
+        })
+}
+
 #[cfg(test)]
 mod tests {
     use crate::{Error, Result};
