+++ src/codec/mod.rs
@@ -25,27 +25,6 @@ pub(crate) mod tcp;
 /// As defined by the spec for both RTU and TCP.
 const MAX_PDU_SIZE: usize = 253;
 
-#[cfg(any(test, feature = "rtu", feature = "tcp"))]
-#[cfg(any(feature = "rtu", feature = "tcp"))]
-pub(crate) async fn disconnect<T, C>(framed: tokio_util::codec::Framed<T, C>) -> std::io::Result<()>
-where
-    T: tokio::io::AsyncWrite + Unpin,
-{
-    use tokio::io::AsyncWriteExt as _;
-
-    framed
-        .into_inner()
-        .shutdown()
-        .await
-        .or_else(|err| match err.kind() {
-            std::io::ErrorKind::NotConnected | std::io::ErrorKind::BrokenPipe => {
-                // Already disconnected.
-                Ok(())
-            }
-            _ => Err(err),
-        })
-}
-
 #[allow(clippy::cast_possible_truncation)]
 fn u16_len(len: usize) -> u16 {
     // This type conversion should always be safe, because either
